refactor: rename files with constants
