+++ src/components/card/constants.ts
