+++ src/components/goods-box/constants.ts
